+++ src/com/example/kiosk/price/Price.java
@@ -1,5 +1,28 @@
 package com.example.kiosk.price;
 
 public class Price {
+    private double totalPrice;
+
+
+    //가격 더하기
+    public void totalPrice(double price){
+        this.totalPrice+=price;
+    }
+
+
+    //총 가격 출력
+    public void printPrice(){
+        System.out.println("[ Total ]");
+        System.out.println("W "+totalPrice);
+    }
+
+    //기존 장바구니 결제 가격 초기화
+    public double totalPriceReset() {
+        return totalPrice=0;
+    }
+
+    public double getTotalPrice() {
+        return totalPrice;
+    }
 
 }
